Switched to signed integers for grid coordinates to facilitate proper support for arithmetic operations.

+++ code/include/allscale/api/user/data/grid.h
@@ -21,7 +21,7 @@ namespace data {
 	// ---------------------------------------------------------------------------------
 
 
-	using coordinate_type = std::size_t;
+	using coordinate_type = std::int64_t;
 
 	template<unsigned Dims>
 	using GridPoint = Vector<coordinate_type,Dims>;
